fix: always make requests to unpkg over HTTPS (#63)
We were making the requests against HTTP which returned a 301 redirect
to the HTTPS protocol. However, this didn't include CORS headers, so,
everything broke. I opened an issue with unpkg but it may not be
possible to fix, so, we should just use HTTPS.

See: https://github.com/unpkg/unpkg.com/issues/22.